aaaaa

+++ scripts/script.js
@@ -1,9 +1,11 @@
-// запрет зума на iOS (не пашет)
-document.addEventListener('touchmove', function (event) {
-    if (event.scale !== 1) { event.preventDefault(); }
-}, false);
+// // запрет зума на iOS (не пашет)
+// document.addEventListener('touchmove', function (event) {
+//     if (event.scale !== 1) { event.preventDefault(); }
+// }, false);
 // отчистка хранилища
-localStorage.clear()
+if(localStorage.length > 1) {
+    localStorage.clear()
+}
 // обработка селекта
 localStorage.setItem('totalPlayers', 3)
 function setPlayers(players) {
